Implemented download dialog & Qt4 support
This commit implements a download dialog to be used inside the
application (removing the need of opening the link in a web browser).

The dialog manages the downloads, gives information about them
(progress, time remaining, size, etc) and prompts the user to close the
app while updating it.

Finally, the commit also provides support for Qt4 by changing the way
the UserAgent header is created.